libnixstore: expose nar size as usize instead of u64
this fixes the build on 32-bit

+++ libnixstore/src/lib.rs
@@ -14,7 +14,7 @@ mod ffi {
         drv: String,
         narhash: String,
         time: i64,
-        size: usize,
+        size: u64,
         refs: Vec<String>,
         sigs: Vec<String>,
         ca: String,
